revert LPS-52538 Use a util class to implement the code for the DeleteAfterTestRunTestRule
